Updated //users, uses hastebin and allows filtering by role.

+++ src/commands/Info/users.js
@@ -1,26 +1,47 @@
-exports.run = function (bot, msg) {
-    let users = msg.guild.members.map(m => `${m.user} ${(m.user.bot ? '**`[BOT]`**' : '')}`).sort();
+function hasRole(member, roleName) {
+    return member.roles.map(role => role.name.toLowerCase()).indexOf(roleName.toLowerCase()) > -1;
+}
 
-    let messages = [];
-    while (users.length > 50) {
-        messages.push(users.splice(0, 50));
+exports.run = function (bot, msg, args) {
+    if (!msg.guild || !msg.guild.members) {
+        throw 'You must run this command from within a server.';
     }
-    messages.push(users);
 
-    msg.edit(':arrows_counterclockwise:');
-    Promise.all(
-        messages.map(group => msg.channel.sendEmbed(
-            bot.utils.embed('', group.join('\n'), [], { footer: false })
-        ))
-    ).then(subMsgs => {
-        msg.delete();
-        subMsgs.forEach(m => m.delete(30000));
-    }).catch(msg.error);
+    let members = msg.guild.members.array().sort((a, b) => a.user.username.localeCompare(b.user.username));
 
+    if (args.length > 0) {
+        members = members.filter(member => hasRole(member, args[0]));
+    }
+
+    if (members.length < 1) {
+        throw 'No members could be found.';
+    }
+
+    msg.delete();
+
+    let users = members.map(m => `${m.user}${(m.user.bot ? ' [BOT]' : '')}`);
+    const body = users.join('\n');
+
+    if (body.length < 2000) {
+        msg.channel.send({
+            embed: bot.utils.embed('', body)
+        }).then(m => m.delete(60000));
+    } else {
+        let raw = members.map(m => `${m.user.username}${m.user.bot ? ' [BOT]' : ''}`).join('\n');
+
+        bot.utils.hastebinUpload(raw).then(res => {
+            let trimmed = body.substr(0, 1500);
+            trimmed = trimmed.slice(0, trimmed.lastIndexOf('\n'));
+
+            msg.channel.send({
+                embed: bot.utils.embed('', trimmed, [{ name: 'Full list', value: res.url }])
+            });
+        }).catch(msg.error);
+    }
 };
 
 exports.info = {
     name: 'users',
-    usage: 'users',
+    usage: 'users [role]',
     description: 'Lists all users on your current server'
 };
